requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
xmltodict==0.13.0
yarg==0.1.9
